Shuffle things around to avoid circ deps
In a perfect world .lib doesn't import from .models...
